Added support for generating multi-protocol struct loading

+++ bpfexamples/ideal.c
@@ -41,7 +41,7 @@ int xdp_filter(struct xdp_md *ctx) {
 
     // OSI 3: Network layer
     // IPv4
-    uint8_t proto4 = -1;     // Proto: 6 TCP, 17 UDP, 1 ICMP, 2 IGMP
+    uint16_t proto4 = -1;     // Proto: 6 TCP, 17 UDP, 1 ICMP, 2 IGMP
     if (proto3 == ETH_P_IP){
         // Return for insufficient data
         if (data + offset + sizeof(*ip) > data_end)
